// lib/api/generated/models/TagPermissionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagPermissionDto = {
  id?: number;
  tagId?: number;
  userId?: number;
  groupId?: number;
  canAccess?: boolean;
  canCreate?: boolean;
  canModify?: boolean;
  canDelete?: boolean;
  createBy?: number;
  createAt?: string;
  updateAt?: string;
  deleteAt?: string;
};